requirements.txt
# data science
wheel
numpy
pandas

# Uncomment if you use sklearn
# scikit-learn

# Install the correct TensorFlow version
tensorflow~=2.13    # For both Intel and Apple silicon



# Google Cloud
google-cloud-storage

python-dotenv
python-multipart

# API
fastapi
uvicorn

#Jupyter
jupyter

#Data sets
datasets

#MLFLOW
mlflow
